+++ server/routes/index.js
@@ -11,12 +11,10 @@ const router = express.Router();
 // Route handlers
 //
 const iotData = require('./iot-data');
-const manualData = require('./manual-data');
 
 //
 // Routes
 //
 router.use('/data/iot-data', iotData);
-router.use('/data/manual-data', manualData);
 
 module.exports = router;